refactor(portal-next): move apis list + search bar to seperate component
