+++ gravitee-apim-portal-webui-next/src/app/catalog/components/apis-list/apis-list.component.html
@@ -1,0 +1,62 @@
+<!--
+
+    Copyright (C) 2025 The Gravitee team (http://gravitee.io)
+
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+            http://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+
+-->
+<div class="api-list__header">
+  <div class="api-list__header__category-info">
+    @if (currentCategory(); as category) {
+      @if (category.name) {
+        <div class="m3-headline-medium">{{ category.name }}</div>
+        @if (category.description) {
+          <div class="m3-label-large">{{ category.description }}</div>
+        }
+      } @else {
+        <div class="m3-headline-medium" i18n="@@allApisLabel">All</div>
+        <div class="m3-label-large" i18n="@@allApisDescription">Explore all available APIs</div>
+      }
+    }
+  </div>
+  <app-search-bar (searchTerm)="onSearchResults($event)" [searchParam]="query()" />
+</div>
+
+@if (apiPaginator$ | async; as apis) {
+  @if (apis.data.length > 0) {
+    <div class="api-list__container" infiniteScroll (scrolled)="loadMoreApis(apis)">
+      @for (api of apis.data; track api.id) {
+        <app-api-card [id]="api.id" [title]="api.title" [version]="api.version" [content]="api.content" [picture]="api.picture" />
+      }
+    </div>
+  } @else if (!loadingPage) {
+    <mat-card id="no-apis" appearance="outlined">
+      <mat-card-content>
+        <div>
+          @if (query() || currentCategory().id) {
+            <p i18n="@@noFilteredApisAvailable">
+              Your search didn't return any APIs. Please try again with different keywords or categories.
+            </p>
+          } @else {
+            <p i18n="@@noApisAvailable">Sorry, there are no APIs listed yet.</p>
+            <p i18n="@@moreApisComingSoon">Our API catalog is currently being updated. More APIs will be available soon.</p>
+          }
+        </div>
+      </mat-card-content>
+    </mat-card>
+  }
+}
+
+@if (loadingPage) {
+  <app-loader />
+}
