+++ gravitee-apim-portal-webui-next/src/app/catalog/components/apis-list/apis-list.component.scss
@@ -1,0 +1,40 @@
+/*
+ * Copyright (C) 2025 The Gravitee team (http://gravitee.io)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+:host {
+  display: flex;
+  flex-flow: column;
+  gap: 32px;
+}
+
+.api-list {
+  &__header {
+    display: flex;
+    place-content: center space-between;
+
+    &__category-info {
+      display: flex;
+      flex-flow: column;
+      justify-content: center;
+      gap: 8px;
+    }
+  }
+
+  &__container {
+    display: grid;
+    gap: 32px;
+    grid-template-columns: repeat(3, minmax(0, 1fr));
+  }
+}
